Converted from .NET to .NET Core. Added base test.
