+++ .gitignore
@@ -3,3 +3,7 @@
 ################################################################################
 
 /.vs
+bin
+obj
+*.user
+*.suo
