+++ PaladinsApiPCL/PaladinsApiPCL-Core.sln
@@ -1,0 +1,28 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.26430.6
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PaladinsApiPCL-Core", "PaladinsApiPCL-Core\PaladinsApiPCL-Core.csproj", "{024E2256-C123-4998-B5E1-D382484EE531}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PaladinsApiPCL-Core.Tests", "PaladinsApiPCL-Core.Tests\PaladinsApiPCL-Core.Tests.csproj", "{375B4F6C-67A3-4202-ADB0-4FF72A6F6A08}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{024E2256-C123-4998-B5E1-D382484EE531}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{024E2256-C123-4998-B5E1-D382484EE531}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{024E2256-C123-4998-B5E1-D382484EE531}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{024E2256-C123-4998-B5E1-D382484EE531}.Release|Any CPU.Build.0 = Release|Any CPU
+		{375B4F6C-67A3-4202-ADB0-4FF72A6F6A08}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{375B4F6C-67A3-4202-ADB0-4FF72A6F6A08}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{375B4F6C-67A3-4202-ADB0-4FF72A6F6A08}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{375B4F6C-67A3-4202-ADB0-4FF72A6F6A08}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
